Removes blocking shutdown for queue drain
This needs to be revisited. If there are connection/pool errors the
queue can't drain, yet there's no way to allow users to resume
shutdown of the server. Ideally we need to only force drain
the queue if there are no errors, print updates in the log so
users don't assume it's frozen, etc.