Moves custom activity data to activities table
Primarily because there's not really any evidence that it's warranted
as a separated table, but also because it's not ideal to manage.

There's a chance the extra inserts can be lost, we have to manually
batch them together which is hacky.

Most critically, not every db (sqlite) supports getGeneratedKeys
on batches so it wasn't working regardless.